Polished breakable logic for the future
